+++ app/src/Header/Header.tsx
@@ -3,16 +3,30 @@ import { LearnMore } from "./LearnMore/LearnMore";
 import { prettyColorPalette } from "../styles/lazyStyles";
 import { words } from "../common/words/words";
 
+/**
+ * Represents a Header component that manages upper level functionality.
+ *
+ * This component allows users to toggle between Spanish and English language modes using a switch.
+ * The component also integrates the `LearnMore` component, passing down the current language mode for
+ * consistent localization. This feature is currently disabled.
+ *
+ * Props:
+ * @param {Object} languageMode - The current language mode settings including text and labels.
+ * @param {Function} setLanguageMode - Function to update the global language mode state.
+ *
+ * State:
+ * @state {boolean} isSpanishMode - Tracks whether Spanish mode is active.
+ * @state {string} languageModeLabel - Displays the current language label (English/Español).
+ *
+ * Behavior:
+ * - Toggles between Spanish and English modes on user interaction.
+ * - Updates language mode globally affecting the entire application scope.
+ * - Dynamically updates labels based on the current state for immediate visual feedback.
+ */
 export const Header = ({ languageMode, setLanguageMode }) => {
   // State for Language mode switch
   const [isSpanishMode, setIsSpanishMode] = useState(false);
   const [languageModeLabel, setLanguageModeLabel] = useState("English");
-
-  // State for Color palette switch
-  const [isHolyGhostModeActive, setIsHolyGhostModeActive] = useState(false);
-  const [colorPaletteLabel, setColorPaletteLabel] = useState(
-    languageMode.buttons["59"]
-  );
 
   // Function to toggle language mode and update label
   const toggleLanguageMode = () => {
@@ -21,18 +35,10 @@ export const Header = ({ languageMode, setLanguageMode }) => {
 
   // Update the language mode label when switch state changes
   useEffect(() => {
-    setLanguageModeLabel(isSpanishMode ? "Español" : "English");
-    setLanguageMode(isSpanishMode ? words["Español"] : words["English"]);
+    setLanguageModeLabel(isSpanishMode ? "Spanish" : "English");
+    setLanguageMode(isSpanishMode ? words["Spanish"] : words["English"]);
   }, [isSpanishMode]);
 
-  // Update the color palette label when switch state changes
-  useEffect(() => {
-    setColorPaletteLabel(
-      isHolyGhostModeActive
-        ? languageMode.buttons["60"]
-        : languageMode.buttons["59"]
-    );
-  }, [isHolyGhostModeActive, languageMode]);
   return (
     <div style={{ color: prettyColorPalette.softYellowGlow }}>
       <LearnMore languageMode={languageMode} />
